+++ routes/api.php
@@ -38,7 +38,9 @@ Route::group(['middleware' => 'auth:api'], function () {
     Route::post('apikey/activatekey/{institution}', [ ApiKeyController::class, 'activateApiKey']); #activate apikey for external offers
 });
 
-
+//test routes for pagination
+Route::get('offer/paginated/{offerCount}', [OfferController::class, 'paginatedOffers']);
+Route::get('list/offer/short/paginated/{offerCount}', [ OfferController::class, 'paginatedReducedOffers']);
 
 // Additional routes that don't require authentication
 Route::get('offer', [ OfferController::class, 'index']);